prototype/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=1.10
httpx>=0.27
python-dotenv>=1.0
